Move Editor Canvas Into UI + UI Improvements
* Increase the default window size from 800x600 to 1024x768.
* Move the drawing canvas in EditorMode to inside the EditorUI where it can
  be better managed with the other widgets it shares the screen with.
* Slightly fix Frame packing bug (with East orientation) that was causing
  right-aligned statusbar items to be partially cropped off-screen. Moved a
  couple statusbar labels in EditorMode to the right.
* Add `Parent()` and `Adopt()` methods to widgets for when they're managed
  by containers like the Frame.
* Add utility functions to UI toolkit for computing a widget's Absolute
  Position and Absolute Rect, by crawling all parent widgets and summing
  them up.
* Add `lib/debugging` package with useful stack tracing utilities.
* Add `make guitest` to launch the program into the GUI Test.
  The command line flag is: `doodle -guitest`
* Console: add a `close` command which returns to the MainScene.
* Initialize the font cache directory (~/.cache/doodle/fonts) but don't
  extract the fonts there yet.

+++ render/sdl/sdl.go
@@ -31,12 +31,12 @@ type Renderer struct {
 }
 
 // New creates the SDL renderer.
-func New(title string, width, height int32) *Renderer {
+func New(title string, width, height int) *Renderer {
 	return &Renderer{
 		events: events.New(),
 		title:  title,
-		width:  width,
-		height: height,
+		width:  int32(width),
+		height: int32(height),
 	}
 }
